Contour tracing - inner and outer contour support
This implementation tracks inner and outer contours plus blob size but
not contour hierarchy as some libraries do.

- Contour data type
- Contour tracing function
- Contour drawing
- Contour sizes
- A test/example in test/Contour.hs (requires as yet unreleased JP 3.2.7)

N.B. At one point there was a bug with contour tracing where it would throw an
exception if:

1. The first pixel of an outer contour was also the first pixel of an inner
   contour.
2. The first pixel after this inner contour was also a first pixel for yet
   another inner contour.

This second condition is more of the symptom - a different exception would have
occurred if the second condition weren't true or slightly altered.

The cited paper was... unclear... that both step 1 _and_ step 2 can
simultaneously apply to the current pixel (not exclusive).  If we detect an
outer contour DO NOT increment the index and loop, instead consider the
possibility we have also detected an inner contour too.

.